move folders around to expose only three top-level directories: client, common, server
