+++ src/client/http/http-request.interface.ts
