+++ src/common/lobby/index.ts
@@ -1,0 +1,1 @@
+export * from "./lobby.interface"
