+++ src/common/lobby/lobby.interface.ts
