--- src/net/client/index.ts
@@ -1,3 +1,0 @@
-export * from './base-web-client'
-export * from './client-packet'
-export * from './http'
