+++ src/server/base-server.ts
@@ -1,9 +1,8 @@
 import express from "express"
 import http from "http"
 import https from "https"
-import { Guard } from "../guard"
-import { BaseLobbyManager } from "../lobby"
-import { Route } from "../route"
+import { BaseLobbyManager } from "./lobby"
+import { Guard, Route } from "./rest"
 
 export abstract class BaseServer {
 
@@ -12,25 +11,20 @@ export abstract class BaseServer {
   constructor(
     public api: express.Express,
     public port: number,
-    public ws: boolean = false,
     public routes: Route[] = [],
     public guards: Guard[] = [],
     private lobbyManager?: BaseLobbyManager
   ) {
-
+    this._configureGuards()
+    this._configureRoutes()
   }
 
-  public onInit(): void {
-    this.configureGuards()
-    this.configureRoutes()
-  }
-
-  private configureRoutes(): void {
+  private _configureRoutes(): void {
     for (let i = 0; i < this.routes.length; i++)
       this.api.use(this.routes[i].path, this.routes[i].router)
   }
 
-  private configureGuards(): void {
+  private _configureGuards(): void {
     for (let i = 0; i < this.guards.length; i++)
       this.api.use(this.guards[i].handler)
   }
@@ -47,9 +41,8 @@ export abstract class BaseServer {
       if (_url.startsWith("/lobby")) {
         const _lobbyId = _request.url.split('/').pop()
         this.lobbyManager?.getLobbies().forEach(_lobby => {
-          if (_lobbyId && +_lobbyId === _lobby.id) {
+          if (_lobbyId && +_lobbyId === _lobby.id)
             _lobby.upgrade(_request, _socket, _head)
-          }
         })
       }
     })
