+++ src/server/lobby/index.ts
@@ -1,5 +1,3 @@
 export * from "./lobby"
 export * from "./lobby-manager"
 export * from "./lobby-manager.interface"
-export * from "./lobby.interface"
-
