+++ src/server/lobby/lobby-manager.interface.ts
