+++ src/server/rest/endpoint/decorators/delete.decorator.ts
