+++ src/server/rest/endpoint/endpoint.interface.ts
