+++ tests/server/test-route.ts
@@ -1,7 +1,6 @@
 import { Syringe } from "@kuroi/syringe"
 import { Request, Response } from "express"
-import { Get } from "../../src/endpoint"
-import { Route } from "../../src/route"
+import { Get, Route } from "../../src/server/rest"
 import { TestGuard } from "./test-guard"
 
 @Syringe.Injectable()
